Set modOrder flag (wrap key data) + err check
This adds error checking (was missing in several places) and sets modOrder flag to true so the bytes wrap around when greater than the mod order for private keys.

@quantumexplorer I assume the bug you reported is probably from setting the 2nd arg in bls.PrivateKeyFromBytes to false when it should probably be true since you are getting them from a random source.

Then it was probably 0 bytes or something so might have "worked" for signing but nothing to be freed. But this is just a guess, can you test this w/these changes and see how it goes? You can also just add the err handling but re-set the flag to false (e.g.: `bls.PrivateKeyFromBytes(privKey, false)`) to see if the error tells you anything.